+++ docs/changelog.rst
@@ -30,6 +30,7 @@ Changelog
 * :feature:`11702` Zerox Base Swaps through the latest settler  will now be properly decoded.
 * :bug:`-` ZKSynclite batch withdrawal on the ethereum side should now be properly decoded.
 * :bug:`-` Claiming AAVE rewards and immediately restaking them will now be properly decoded.
+* :bug:`-` The rare case of balancer v2 swaps with repeated swap logs for a single events is now decoded properly.
 
 * :release:`1.42.0 <2025-02-20>`
 * :feature:`-` Octant v2 migration is now supported and GLM locked in v2 detected.